move forms into respective type category folders
